backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
scikit-learn
chromadb
uvloop
httpx
pydantic
llama-cpp-python
